+++ testing/ringbuffer.h
@@ -7,9 +7,9 @@ class RingBuffer {
     explicit RingBuffer(int size_ = 128); //128 * 1024, 0x20000
     ~RingBuffer();
     size_t getNumUnread();
-    unsigned int getReadPosition(); 
-    unsigned int getWritePosition();
-    unsigned long long getNumProcessed();
+    size_t getReadPosition(); 
+    size_t getWritePosition();
+    size_t getNumProcessed();
     uint8_t readNext();
     size_t populate(char* buff, size_t size, size_t exclude_from_front = 0);
     void setRollback();
@@ -19,11 +19,11 @@ class RingBuffer {
   private:
 
     char* buffer_;
-    const int size_;
-    unsigned int read_position_;
-    unsigned int write_position_;
-    unsigned int rollback_;
-    unsigned long long data_processed_;
+    const size_t size_;
+    size_t read_position_;
+    size_t write_position_;
+    size_t rollback_;
+    size_t data_processed_;
     size_t unread_;
     size_t ring_mask_;
 };
